+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 ==========
 
+Version 1.0.3 *(2024-03-12)*
+----------------------------
+
+* Fixed month view issue on Google Pixel 8 Pro.
+* Fixed event color dots on monthly and daily view.
+* Fixed incorrect timezone when import ICS files.
+* Highlight weekends in print mode.
+* Updated holidays for some countries.
+* Added some translations.
+
 Version 1.0.2 *(2024-01-02)*
 ----------------------------